% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howl.R
\name{howl}
\alias{howl}
\title{Calculate power of a  wolf howl}
\usage{
howl(age, length)
}
\arguments{
\item{age}{Age of wolf}

\item{length}{Length in seconds wolf howl is}
}
\value{
Returns the power of a wolf howl
}
\description{
Calculate power of a  wolf howl
}
\examples{
howl (3,8)
}
